pyproject.toml
[tool.poetry]
name = "uc"
version = "0.1.0"
description = ""
authors = ["Patrik Persson <patrik95195@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^8.1.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
